Remove console. Fix tests.
