+++ qb/test/collections.test.ts
@@ -1,6 +1,5 @@
 import {Client, $} from "edgedb";
 import e, {$infer} from "../dbschema/edgeql-js";
-
 import type {$VersionStageλEnum} from "../dbschema/edgeql-js/modules/sys";
 import {tc} from "./setupTeardown";
